requirements.txt
pandas
requests
SQLAlchemy
python-dotenv
mysql-connector-python
matplotlib
seaborn
ipython-sql
sqlalchemy
prettytable==3.9.0